+++ zcash_client_backend/CHANGELOG.md
@@ -63,7 +63,8 @@ workspace.
     that would otherwise only be recoverable using the wallet's internal
     IVK.
 - `zcash_client_backend::data_api::WalletRead` has added method `get_received_outputs`.
-- `zcash_client_backend::proposal::ProposalError` has added variant `PaymentAmountMissing`
+- `zcash_client_backend::proposal::ProposalError` has added variants `PaymentAmountMissing`
+  and `IncompatibleTxVersion`
 - The associated type `zcash_client_backend::fees::ChangeStrategy::MetaSource` is now
   bounded on the newly added `MetaSource` type instead of
   `zcash_client_backend::data_api::InputSource`.